I expect that after installing this plugin in the target Intellij idea, it can have a configuration page in Intellij-idea\Setting\Tool\ section.
The configuration page contains 2 tab pages.

First tab page is for showing a list of the data, let's name them rules, exist in the JSON file which is cached by the Intellij idea (IDE).
Second tab page for importing a JSON file or show the one that user imported before, like the file path. And a button to allow user to export the list of rules as JSON file.

The first tab page that shows list of rules, has a plus button to add new rule to the list and a disable button to deactivate the selected rule.

The sample JSON file content is like this:

{
  "rules": [
    {
      "name": "rule-name-1",
      "description": "rule-description-1",
      "active": true,
      "works_over": {
        "method": true,
        "class": false
      },
      "priority": 1,
      "priority-meaning": "normal",
      "context": "rule context 1",
      "url": "http://www.test1.com/rule/82645"
    },
    {
      "name": "rule-name-2",
      "description": "rule-description-2",
      "active": true,
      "works_over": {
        "method": false,
        "class": true
      },
      "priority": 2,
      "priority-meaning": "medium",
      "context": "rule context 2",
      "url": "http://www.test1.com/rule/82385"
    }
  ]
}

I expect to import the rule JSON file and see the lust of rules.
I would like to double click one of the rule and in the popup window it allow me to edit the rule's data. Same, if I click the add new rule button, it allow me to add rule data in the popup window ad it add the new rule to the list of rules.